When creating a file in the B-tree, don't nuke its previous data.
This is important so that we can add files to a directory
before we add the directory. And that, in turn, is important
because it makes it easier to write upper levels of the backup
application. There is no longer a need for it to carefully
do things in a particular order.