Change put_chunk to use random chunk ids.
Previously, it started iterating from 0 to find
the first free id, and that's an obviously bad
way of doing it.

Using random ids means we do not need to keep store
the last used id anywhere, which simplifies things.
We do need to check if the id is in use, but with
a 64-bit address space, collisions should be rare
enough that it won't cause performance problems.