Change how directory contents are stored, for higher speed.
Previously, the contents of a directory was stored using
key(dirname, DIR_CONTENTS, counter)=namehash, where counter
was increased for each new entry, and namehash was the 64-bit
hash of the child's pathname. The problem this has is that
when listing the contents of a directory, we have to look
up the pathname for each child, and that's pretty slow.

Now we store the contents of the directory as 
key(dirname, DIR_CONTENTS, namehash)=basename, so that listdir
can be done with a single range query.

This does not handle hash collisions, but that'll be doable,
we just need to make sure we find and store the right namehash
when inserting the value, and possibly when accessing the
child as well.