Move SftpFS into a plugin and start getting it to work.
Add test-sftpfs to test it. Since it requires access to a
working sftp server, can't run the tests using normal
unit tests, unless I create a dummy sftp server, and I
don't want to do that, so this is good enough for now.

Also make the plugins directory be a module so that I can
import from the sftp plugin in test-sftpfs.