Fix test for sparse files.
We can't necessarily get the wholes in the file in the exact same places, so we
can't compare the metadata the same way we normally do. Instead, we just check
that the file contents look identical (i.e., the right number of zeroes, even if
they may or may not come from wholes).

This does not truly check that restore puts wholes in the restored file, but
it's good enough for now.