Improve speed of pruning.
The old code had this problem: "removes entries from the 
'items' list in a way that can be O(n^2) in the number of 
items (if it deletes items[0] each time until the list is 
empty)."

The new code should be linear.

Problem found and fixed by Jeff Epler. Thanks!