make verify-notices-changes test handle whole-second timestamps
We need whole-second timestamps for sftp, and if the test is run
suitably, then the timestamp change made by touch wouldn't actually
cause a change. Forcing touch to use a very old timestamp fixes this.

Anyone running their computer with the clock set to 1970 will need
to fix that first.