During lock testing, handle obnam restore failures gracefully
Because of how Obnam deals with locking (it only locks for writes),
read-only operations such as restore can fail because a backup is
happening concurrently. We deal with this by re-trying the restore
until it succeeds (i.e., obnam does not crash). There's a chance
of infinite looping if the crash happens for another reason, but
that's for the human watching the test to deal with.