Fix problem with adding clients during backup
Found a problem where client-1 adds itself to the
repository, and then client-2 does the same, but
client-2 had already read the B-tree into memory
before client-1 added itself, and so when client-2
saved the modified B-tree, it accidentally overwrote
the changes client-1 had made, resulting in no
client-1 in the repository. Oops.