Order work items so the list of items clears faster
Previously, all new work items were added to the end
of the queue. This resulted in about half the items
being unprocessed by the time the last item was added.
We now add new items from an item being processed
to the other end of the queue (but still in the right
relative order), which solves the problem.