Make Obnam handle value-less extended attributes
What happens:

* llistxattr returns a list of names
* lgetxattr for one or more names returns ENODATA (ENOATTR)

In other words, the name exists, but has no value. This is
different from it having an empty value. This happens at least
sometimes on a btrfs filesystem, but in my opinion should not
be happening at all. However, since it does, I have changed
Obnam to deal with this by pretending the name does not exist.
It logs a warning, but otherwise it is as if the name had not
been returned by llistxattr.

It would not be acceptable to pretend the value is empty, I
think. An empty value may have some significance that is
different from a value-less name. (I'm happy to change this
if it turns out to be the wrong decision.)

I was unable to reproduce this until Vladimir made me a btrfs
disk image, after which this was easy to debug.

While debugging this, I also realised that lgetxattr in LocalFS
creates an OSError wrongly. Fixed that as well.

Reported-by: Vladimir Elisseev