Put work items from current one into queue in right place
If we have a queue [A, B, C], and we start processing A,
and A generates A' and A'', we want to process them before
B. Thus we put A' and A'' to the head of the queue.

We also (now) remove the head of the queue rather than the tail.
The old code tried to do all of these things, but was
majorly confused.