Fix format 6 to not create spurious clone generation
The silly way in which the ClientMetadataTree works, it wants "changes
to be started", and we were doing that in a way that also starts a new
generation, even though we were only removing a generation.

Also, change chunk index operations to take a client name, instead
of client id, since the client id is a concept that only exists in
these two API calls, and that's silly.