Add test for multiple clients sharing a repository
As part of this, make every repository operation step require an
explicit client name. The v6 reference formats have a client
name embedded that is the hostname of my laptop, so make those tests
use that name. This fixes what would not have run previously on any
system that isn't named havelock. Oops.