Remove cmdtest that is useless
This verifies that we can backup a symlink as root. But what we do
with that is surprising: most users seem to expect the directory the
symlink points at to be backed up, not the symlink itself (or not just
the symlink). Thus, I'm removing the test since it's not useful.

Later, we'll need to figure out what we actually do with non-directory
roots, and write tests for those.