Fix file restore, add xattr value change test
The single file restore test had a bug, in that it was depending on
various test steps to execute fast enough that the wall clock did not
advance to the next second. This has been fixed by allowing manifests
to be done on single files instead of just directories.

Add a test for an extended attribute value changing, and that being
the only change in the live data. The backup needs to notice that it
has changed. This was intended to be the original change in this
commit, but I ran into the test bug above, so I had to fix that too.
I was too lazy to detangle the two sets of changes for commits.