Make exclude-caches scenario clearer, robust
Originally, we would create L/wanted, then do a manifest, then create
L/cache, then backup, restore, verify. This would sometimes fail, if
the full-second timestamp of L would change when L/cache was created.
So we (well, I) changed it so that we fiddled with the manifest just
before making a backup, but this was weird and confusing and would
still sometimes fail. This should now be much clearer and therefore
more robust.