Fix parsing of filtered files
This bug was originally triggered by the test suite, because the
error message in the code had a typo. While looking for that I noticed
that the code that was parsing the code seemed fishy.

This is now cleaner. First of all, I split the error messages for
"unknown tag" and "no tag". Then I rewrote the loop that parses
so it is robust against error situations. Then I tweaked the yarn
test so it now works.