Rewrite read() to fix it, simplify and pessimise it
The previous code was broken: it assumed that most chunks (all but the
last one in a file) were of the same size, and that's not a valid
assumption. It is almost always true, but the user may change the
chunk size setting for any backup run, and then the assumption is no
longer true.

The new code is much simpler. It is also, however, much slower, since
it has to read through everything. We can optimise this, later, but
caching the chunk sizes, and even later by making it possible to find
the length of a chunk without downloading it from the repository.