Live data generation progress; simplify restores
If we restore all the data, we need space for an extra copy of the
live data. However, for backup performance purposes, we're OK with
with just measuring the speed of reading from the repository. Thus, we
use "obnam verify" instead of "obnam restore", and tar to /dev/null
instead of cp to restore from a FUSE mount.