Bug fix: change FUSE read to read right amount of data
The actual bug was that when we extracted data from the chunk we were
processing at the time, we used "data[start:n]", where n is the number
of bytes we wanted to extract from this chunk. This is clearly wrong:
the slice end is an offset, not a length to be added to the start.
Thus, "data[start:start+n]" is the fix.

In addition, a small performance silliness is removed by fixing the
check for whether we should be extracting data from a chunk: if the
chunk ends at the start of the offset in the file from where data
should be returned, it should clearly not be considered.

Clarify code a bit as well, which was useful for me to understand
what's going on.