Apply patch from Valery to fix FUSE read
Patch by Valery Yundin. He gave the following recipe to verify it
works:

  * set "ObnamFuseFile.keep_cache" to False
  * backup
  * mount
  * diff -r the original and via FUSE

Before the patch this failed for me if the original file was large
enough, on the order of two to a hundred megs. It works after the
patch.

Someday I will have to really write a test suite for this.