Implement support for listing a generation in kdirstat.cache format
This implements an ls variant whose output format is compatible with the
kdirstat cache format. The best reference I can find for this is from an old fork:
https://github.com/thomas-joiner/k4dirstat/blob/master/kdirstat/cache-file-format.txt

Recent versions of k4dirstat appear to be able to read the files produced just
fine, although this document is not present in their source AFAICT.

I find k4dirstat to be an excellent way to visualise what is included in a
backup, e.g. for the purposes of excluding file types or marking directories
with a CACHEDIR.TAG. My previous solution involved post-processing the output
of ls which require a ssh to the machine with the files (for stat() purposes).
This approach pulls all of the data out of the obnam repo, so is much faster,
less hacky and gets the status of the actual generation and not the current
files etc.