Backup metadata before backing up file content
This is in addition to backing it up after the file content. It's
necessary to back it up afterwards so that the whole file checksum is
set. Backing metadata up before the file content is necessary so that
the file type is set before a checkpoint generation happens, due to
changes required by Green Albatross.