Don't see if root is file, it isn't
We do not, in fact, currently support the live data root being
a non-directory. The code removed here assumes it can be (it can't,
neither LocalFS nor the SFTP plugin support it), and (sometimes)
gets the check for whether it is wrong, resulting in backing up the
parent directory instead, breaking things and confusing people.