kdirstat: Handled named pipes (fifos)
For some reason this was missing from the mode_str lookup, it is the only type
listed on
https://github.com/thomas-joiner/k4dirstat/blob/master/kdirstat/cache-file-format.txt
which wasn't handled.

For unknown file types emit a commented "#UNHANDLED ..." line rather than
crashing with a backtrace.

Add a yarn for this scenario, although I've not been able to run it due to
"ASSUMING extended attributes are allowed for users".

Real world use has been tested by Florian Geyer,
http://listmaster.pepperfish.net/pipermail/obnam-support-obnam.org/2015-September/003836.html

Cc: Florian Geyer <blueice@fobos.de>